+++ src/screens/Calendar/PromoCards.tsx
@@ -7,10 +7,30 @@ import { useAnalytics } from "@/hooks/useAnalytics";
 import { useSettings } from "@/hooks/useSettings";
 import { useNavigation } from "@react-navigation/native";
 import dayjs from "dayjs";
-import React, { ReactElement } from "react";
+import React, { ReactElement, useEffect, useState } from "react";
 import { View } from "react-native";
 import useColors from "../../hooks/useColors";
 import { useLogState } from "../../hooks/useLogs";
+import * as rssParser from 'react-native-rss-parser';
+import * as WebBrowser from 'expo-web-browser';
+
+type RssItem = {
+  title: string,
+  links: {
+    url: string,
+    rel: string,
+  }[],
+  description: string,
+  id: string,
+  authors: string[],
+  categories: string[],
+  published: string,
+  enclosures: string[],
+  itunes: {
+    authors: string[],
+  }
+  slug: string,
+}
 
 export const PromoCards = () => {
   const navigation = useNavigation();
@@ -26,8 +46,29 @@ export const PromoCards = () => {
 
   const hasMonthPromo = isBeginningOfMonth && statisticsUnlocked && !hasActionDone(MONTH_REPORT_SLUG)
   const hasYearPromo = enoughtLogsForYearPromo && isDecember && statisticsUnlocked && !hasActionDone(YEAR_REPORT_SLUG)
-  const hasEmotionTrackingPromo = logState.items.length > 4 && !hasActionDone('promo_emotions_tracking_closed')
   const hasSleepTrackingPromo = logState.items.length > 4 && !hasActionDone('promo_sleep_tracking_closed')
+
+  const [mostRecentRssItem, setMostRecentRssItem] = useState<RssItem | null>(null)
+
+  const hasMostRecentRssItem = !!mostRecentRssItem && !hasActionDone(mostRecentRssItem.slug)
+
+  useEffect(() => {
+    fetch('https://pixy.hellonext.co/rss/changelog.xml')
+      .then(response => response.text())
+      .then(str => rssParser.parse(str))
+      .then(rss => {
+        const items = rss.items
+          .filter(item => dayjs(item.published).isAfter('2023-01-09'))
+          .map(item => ({
+            ...item,
+            slug: item.id.replace(/[^a-z0-9]/gi, '_').toLowerCase(),
+          }))
+
+        if (items.length !== 0) {
+          setMostRecentRssItem(items[0])
+        }
+      })
+  }, [])
 
   const promoCards: ReactElement[] = []
 
@@ -58,6 +99,21 @@ export const PromoCards = () => {
         onPress={() => {
           analytics.track('promo_sleep_tracking_clicked')
           navigation.navigate("Steps");
+        }}
+      />
+    )
+  }
+
+  if (hasMostRecentRssItem) {
+    promoCards.push(
+      <PromoCard
+        colorName="pink"
+        slug={mostRecentRssItem.slug}
+        subtitle={t('new_release')}
+        title={mostRecentRssItem.title}
+        onPress={() => {
+          analytics.track('promo_changelog_clicked')
+          WebBrowser.openBrowserAsync(mostRecentRssItem.id);
         }}
       />
     )